refactor(vlm): define llava_ov_mapstyle_dataloader in llava_ov_experiment.py via deepcopy
pre_exp012_llava_ov_mapstyle_dataloader differed from pre_exp012_llava_ov only in
job.name, the dataloader_state callback (distributor_type="cosmos_dataloader"),
checkpoint.save_iter, and the dataloader (streaming IterableDistributor vs
map-style MapDistributor + num_workers=0). Derive it by deep-copying
pre_exp012_llava_ov and overriding just those fields; move get_llava_ov_map
alongside get_llava_ov_streaming. Delete the standalone module (VLM experiments
auto-register via import_all_modules_from_package). Experiment name and TOML
unchanged.
